pyproject.toml
[tool.poetry]
name = "camera-mlx90640"
version = "0.1.0"
description = ""
authors = ["Will Dickson <wbd@caltech.edu>"]
readme = "README.md"
packages = [{include = "camera_mlx90640"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pyserial = "^3.5"
numpy = "^1.25.0"
matplotlib = "^3.7.1"
scipy = "^1.11.2"
click = "^8.1.7"

[tool.poetry.scripts]
mlx90640-live = 'camera_mlx90640.live_view_cli:cli'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
